requirements.txt
openai>=1.30.0,<2.0.0
google-generativeai>=0.3.1
python-dotenv>=1.0.0
requests>=2.31.0
